refactor location of atmosphere profile functions
- moved atmosphere funcs to `atmosphere.py` (imported into the
instrument namespace)
- moved camera util funcs into `camera.py` (also imported to instrument)
- update tools that use them
