+++ ctapipe/instrument/__init__.py
@@ -1,5 +1,4 @@
-from .camera import CameraGeometry
+from .camera import *
+from .atmosphere import *
 
-__all__ = ['CameraGeometry',]
 
-from .utils import *
